requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
bs4==0.0.1
click==8.1.3
Django==4.1.7
django-ckeditor==6.5.1
django-js-asset==2.0.0
djangorestframework==3.14.0
importlib-metadata==6.0.0
joblib==1.2.0
Markdown==3.4.1
nltk==3.8.1
Pillow==9.4.0
pytz==2022.7.1
regex==2022.10.31
soupsieve==2.4
sqlparse==0.4.3
tqdm==4.65.0
zipp==3.15.0
